api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
transformers
torch
python-docx
python-dotenv
